infrastructure/requirements.txt
pandas==2.2.3
seaborn
matplotlib.pyplot
